Migrate ScheduleCompetitionEventsJob
