+++ server/src/api/jobs/instances/ScheduleCompetitionEventsJob.ts
@@ -1,20 +1,5 @@
-import prisma from '../../../prisma';
-import { EventPeriodDelay } from '../../services/external/discord.service';
-import * as competitionEvents from '../../modules/competitions/competition.events';
+import { ScheduleCompetitionEventsJob as NewScheduleCompetitionEventsJob } from '../../../jobs/instances/ScheduleCompetitionEventsJob';
 import { JobType, JobDefinition } from '../job.types';
-
-// Since the cronjob runs at every minute (at 00 seconds) and most competitions start at 00 seconds
-// it is prudent to add a safety gap so that we search dates from X:55 to X+1:55 instead of always at 00
-const SAFETY_GAP = 5_000;
-
-// How often this job is executed (once every minute)
-const EXECUTION_FREQUENCY = 60_000;
-
-// 6h, 5min, now
-const START_TIME_INTERVALS = [360, 5, 0];
-
-// 12h, 2h, 30min, now
-const END_TIME_INTERVALS = [720, 120, 30, 0];
 
 class ScheduleCompetitionEventsJob implements JobDefinition<unknown> {
   type: JobType;
@@ -24,83 +9,10 @@ class ScheduleCompetitionEventsJob implements JobDefinition<unknown> {
   }
 
   async execute() {
-    // Schedule "starting" and "started" events for each interval
-    for (const start of START_TIME_INTERVALS) {
-      await scheduleStarting(start * 60 * 1000);
-    }
-
-    // Schedule "ending" and "ended" events for each interval
-    for (const end of END_TIME_INTERVALS) {
-      await scheduleEnding(end * 60 * 1000);
-    }
+    // We're migrating to a new job manager, but jobs for the current one are still in-queue
+    // so to prevent duplicating code, just use the old job manager to execute the job on the new one
+    new NewScheduleCompetitionEventsJob().execute();
   }
-}
-
-async function scheduleStarting(delayMs: number): Promise<void> {
-  const startSearchDate = new Date(Date.now() - SAFETY_GAP + delayMs);
-  const endSearchDate = new Date(startSearchDate.getTime() + EXECUTION_FREQUENCY);
-
-  const competitionsStarting = await prisma.competition.findMany({
-    where: {
-      startsAt: {
-        gte: startSearchDate,
-        lte: endSearchDate
-      }
-    }
-  });
-
-  competitionsStarting.forEach((c, index) => {
-    const eventDelay = Math.max(0, c.startsAt.getTime() - delayMs - Date.now());
-
-    setTimeout(
-      () => {
-        // If competition is starting in < 1min, schedule the "started" event instead
-        if (delayMs === 0) {
-          competitionEvents.onCompetitionStarted(c);
-        } else {
-          competitionEvents.onCompetitionStarting(c, getEventPeriodDelay(delayMs));
-        }
-        // stagger each event by 500ms to avoid overloading the database
-      },
-      eventDelay + index * 500
-    );
-  });
-}
-
-async function scheduleEnding(delayMs: number): Promise<void> {
-  const startSearchDate = new Date(Date.now() - SAFETY_GAP + delayMs);
-  const endSearchDate = new Date(startSearchDate.getTime() + EXECUTION_FREQUENCY);
-
-  const competitionsEnding = await prisma.competition.findMany({
-    where: {
-      endsAt: {
-        gte: startSearchDate,
-        lte: endSearchDate
-      }
-    }
-  });
-
-  competitionsEnding.forEach((c, index) => {
-    const eventDelay = Math.max(0, c.endsAt.getTime() - delayMs - Date.now());
-
-    setTimeout(
-      () => {
-        // If competition is ending in < 1min, schedule the "ended" event instead
-        if (delayMs === 0) {
-          competitionEvents.onCompetitionEnded(c);
-        } else {
-          competitionEvents.onCompetitionEnding(c, getEventPeriodDelay(delayMs));
-        }
-        // stagger each event by 500ms to avoid overloading the database
-      },
-      eventDelay + index * 500
-    );
-  });
-}
-
-function getEventPeriodDelay(delayMs: number): EventPeriodDelay {
-  const minutes = delayMs / 1000 / 60;
-  return minutes < 60 ? { minutes } : { hours: minutes / 60 };
 }
 
 export default new ScheduleCompetitionEventsJob();
